move fillExtra to node.go
