+++ node_linux.go
@@ -1,10 +1,7 @@
 package restic
 
 import (
-	"fmt"
 	"os"
-	"os/user"
-	"strconv"
 	"syscall"
 	"time"
 )
@@ -17,43 +14,13 @@ func (node *Node) OpenForReading() (*os.File, error) {
 	return file, err
 }
 
-func (node *Node) fillExtra(path string, fi os.FileInfo) error {
-	stat, ok := fi.Sys().(*syscall.Stat_t)
-	if !ok {
-		return nil
-	}
-
+func (node *Node) fillTimes(stat *syscall.Stat_t) {
 	node.ChangeTime = time.Unix(stat.Ctim.Unix())
 	node.AccessTime = time.Unix(stat.Atim.Unix())
-	node.UID = stat.Uid
-	node.GID = stat.Gid
+}
 
-	if u, err := user.LookupId(strconv.Itoa(int(stat.Uid))); err == nil {
-		node.User = u.Username
-	}
-
-	node.Inode = stat.Ino
-
-	var err error
-
-	switch node.Type {
-	case "file":
-		node.Size = uint64(stat.Size)
-		node.Links = uint64(stat.Nlink)
-	case "dir":
-	case "symlink":
-		node.LinkTarget, err = os.Readlink(path)
-	case "dev":
-		node.Device = stat.Rdev
-	case "chardev":
-		node.Device = stat.Rdev
-	case "fifo":
-	case "socket":
-	default:
-		err = fmt.Errorf("invalid node type %q", node.Type)
-	}
-
-	return err
+func (node *Node) fillDevice(stat *syscall.Stat_t) {
+	node.Device = stat.Rdev
 }
 
 func (node *Node) createDevAt(path string) error {
